Project cleanup

+++ include/Config.h
@@ -11,5 +11,5 @@ namespace Config
 
     const float dtDebug = PhysicsManager::dt / 4.0;
 
-    const float targetFPS = 200;
+    const int targetFPS = 200;
 }
